Refactor about buttons section
